Infer the project name from the project
Create a new action that infers the project name from the project, as long as there is a target that matches the project name.
This action allows to be provided a project_name in case the inferred name should not be used.
Taking advantage of this action, actions that inferred the project name by their own, now extend this new one, delegating this behavior.
Also, properties from .env.default have been moved to Matchfile, since they have to be there in case match is run from terminal.